+++ Changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix abi3 wheel build when no Python interpreters found in [#1072](https://github.com/PyO3/maturin/pull/1072)
 * Add `zig ar` support in [#1073](https://github.com/PyO3/maturin/pull/1073)
 * Fix sdist build for optional path dependencies in [#1084](https://github.com/PyO3/maturin/pull/1084)
+* auditwheel: find dylibs in Cargo target directory in [#1092](https://github.com/PyO3/maturin/pull/1092)
 
 ## [0.13.2] - 2022-08-14